Runtime/socket: fix size_t/int problem
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@30275 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/socket.cpp
@@ -630,7 +630,7 @@ RTDECL(int) RTSocketSgWrite(RTSOCKET hSocket, PCRTSGBUF pSgBuf)
         int cbBuf = 0;
         for (unsigned i = 0; i < pSgBuf->cSeg; i++)
         {
-            cbBuf += pSgBuf->pcaSeg[i].cbSeg;
+            cbBuf += (int)pSgBuf->pcaSeg[i].cbSeg;
             AssertBreakStmt(cbBuf, rc = VERR_BUFFER_OVERFLOW);
         }
 
@@ -640,7 +640,7 @@ RTDECL(int) RTSocketSgWrite(RTSOCKET hSocket, PCRTSGBUF pSgBuf)
         for (unsigned i = 0; i < pSgBuf->cSeg; i++)
         {
             memcpy(p, pSgBuf->pcaSeg[i].pvSeg, pSgBuf->pcaSeg[i].cbSeg);
-            p += pSgBuf->pcaSeg[i].cbSeg;
+            p += (int)pSgBuf->pcaSeg[i].cbSeg;
         }
 
         ssize_t cbWritten = send(pThis->hNative, (const char *)pBuf, cbBuf, MSG_NOSIGNAL);
